Development

+++ TransportGlobal/TransportGlobalAPI/src/TransportGlobal.Application/CQRSs/TransporterContextCQRSs/CommandDeleteVehicle/DeleteVehicleCommandHandler.cs
@@ -28,6 +28,8 @@ namespace TransportGlobal.Application.CQRSs.TransporterContextCQRSs.CommandDelet
             UserEntity userEntity = _userRepository.GetByID(userID) ?? throw new ClientSideException(ExceptionConstants.NotFoundUser);
 
             VehicleEntity vehicleEntity = _vehicleRepository.GetByID(request.ID) ?? throw new ClientSideException(ExceptionConstants.NotFoundVehicle);
+            if (vehicleEntity.IsDeleted) throw new ClientSideException(ExceptionConstants.NotFoundVehicle);
+
             if (vehicleEntity.CompanyID != userEntity.ActiveCompany?.ID) return Task.FromResult(new DeleteVehicleCommandResponse(ResponseConstants.NotVehicleOwner));
 
             if (vehicleEntity.Employees.Count > 0)
@@ -35,15 +37,11 @@ namespace TransportGlobal.Application.CQRSs.TransporterContextCQRSs.CommandDelet
                 foreach (EmployeeEntity employee in vehicleEntity.Employees)
                 {
                     employee.VehicleID = null;
-                    _employeeRepository.Update(employee);
                 }
-
-                // TODO: Test sürecinde kodun gerekli olup olmadığına bakılacak
-                int updateEffectedRows = _employeeRepository.SaveChanges();
-                if (updateEffectedRows == 0) return Task.FromResult(new DeleteVehicleCommandResponse(ResponseConstants.DeleteFailed));
             }
 
-            _vehicleRepository.Delete(vehicleEntity);
+            vehicleEntity.IsDeleted = true;
+            _vehicleRepository.Update(vehicleEntity);
 
             int effectedRows = _vehicleRepository.SaveChanges();
             if (effectedRows == 0) return Task.FromResult(new DeleteVehicleCommandResponse(ResponseConstants.DeleteFailed));
